Initial commit of the new console view.
The console view is a new StatusResource that displays
a developer-oriented view. It makes it easy to know
what was the result of a given change on all the builders.

This is still a work in progress, but this version
should be working. It supports only integer-based
SCM at this point (svn, perforce, etc..., not git).